Create buildPolygonCells
In addition to the routine that builds cells as centroids, this routine builds the set of cells for tiles as individual polygons. This is used to provide improved function when AusPIX DGGS enabling of data sets consisting of point or line geometries. Used with the geoPandas method of building crosswalk tables.